+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Bob versions changelog
 #### Changed
 - Use cargo workspace to declare dependencies to avoid their duplication (#821)
 - Record timestamp is now passed to Pearl level and used to find newest record in get and exist functions (#708)
+- Save gRPC error when parsing status (#842)
 
 #### Fixed
 - Fix missing alien records due to multiple groups (#806)